requirements.txt
requests==2.22.0
pycrypto==2.6.1
pyprnt==2.3.4
Flask==1.1.1
coverage==4.5.4
